requirements.txt
numpy
rawpy
opencv-python-headless
azure-cognitiveservices-vision-computervision